+++ vulp/exceptions/tests/observer_error_test.cpp
@@ -9,9 +9,9 @@
 #include <vector>
 
 #include "gtest/gtest.h"
-#include "vulp/observation/ObserverError.h"
+#include "vulp/exceptions/ObserverError.h"
 
-namespace vulp::observation::tests {
+namespace vulp::exceptions::tests {
 
 TEST(ObserverError, NineteenSixtyWho) {
   const std::string prefix = "nineteen_sixty";
@@ -22,4 +22,4 @@ TEST(ObserverError, NineteenSixtyWho) {
   ASSERT_NE(msg.find(key), std::string::npos);
 }
 
-}  // namespace vulp::observation::tests
+}  // namespace vulp::exceptions::tests