+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 - BulletInterface: Add a new parameter to allow additional URDFs.
 
+### Changed
+
+- Move exceptions to a dedicated namespace
+
 ### Removed
 
 - Bazel: dependencies based on ``pip_parse`` from ``rules_python``